Andrew Cann wrote:
Here's a very small patch that changes one of the arguments to
GNUNET_CONFIGURATION_get_value_choice. C++ doesn't allow you to cast
from a (char **) to a (const char **) because that enables overwriting
read-only memory. This patch makes the argument "more const" (which is
always a good thing) and makes it possible to call it from C++ passing
in a (char **).

https://canndrew.org/misc/change_GNUNET_CONFIGURATION_get_value_choice_argument.diff



git-svn-id: https://gnunet.org/svn/gnunet@29814 140774ce-b5e7-0310-ab8b-a85725594a96